refactor: command 를 domain 패키지로 이동
