+++ src/main/java/chess/domain/command/Move.java
@@ -1,4 +1,4 @@
-package chess.command;
+package chess.domain.command;
 
 import java.util.List;
