+++ src/main/java/chess/domain/command/Start.java
@@ -1,4 +1,4 @@
-package chess.command;
+package chess.domain.command;
 
 import java.util.List;
